Implement Volume-Based Carrying and Resource Cleanup Plans
- Economy & Items (item.rs, resource_catalog.rs, core.ron):
  - Added 'unit_volume_ml' to ResourceDef and Item.
  - Populated volume data for all core resources (e.g., Grain 450ml, Timber 35L).
  - Transitioned storage and carrying capacity logic to be volume-aware.
- Carrying System (carry.rs, person.rs, animals.rs, vehicle.rs):
  - Replaced hard-coded 'HAND_QTY_CAP' with dynamic volume-based limits.
  - Implemented 'HAND_SMALL_VOLUME_ML' (2L) and 'HAND_OVERSIZED_VOLUME_ML' (100L).
  - Workers now carry small items (fruit, seed) by volume and bulky items (logs, stone) by weight/volume envelope.
  - Updated PackAnimal and Vehicle capacities to use volume (ml).
- AI & Simulation:
  - Updated HTN dispatchers and executors (Gather, Dig, Produce) to respect volume-constrained carrying.
  - Refined nomad pool seeding and faction storage systems.
- UI:
  - Updated Hover and Inspector panels to display resource volume.
  - Improved Vehicle Designer validation for volume-based cargo bays.
- Documentation & Planning:
  - Added 'plans/loose-resource-cleanup.md' and 'plans/neolithic-hearth-fix.md'.
  - Updated 'CLAUDE.md' and 'TODO.md' with new carrying mandates.
- Total tests: 1073/1073 passing.

+++ src/simulation/faction.rs
@@ -1982,7 +1982,7 @@ pub fn release_reservation(
     ai: &mut crate::simulation::person::PersonAI,
 ) {
     if let Some(resource_id) = ai.reserved_resource {
-        reservations.sub(ai.reserved_tile, resource_id, ai.reserved_qty as u32);
+        reservations.sub(ai.reserved_tile, resource_id, ai.reserved_qty);
     }
     ai.reserved_resource = None;
     ai.reserved_qty = 0;
